Studium/backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
boto3==1.35.76
botocore==1.35.76
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorclass==2.2.2
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.3
daphne==4.1.2
Django==4.2.18
django-cors-headers==4.4.0
django-csp==4.0
django-ratelimit==4.1.0
django-storages==1.14.6
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
jmespath==1.0.1
kombu==5.4.2
libmagic==1.0
msgpack==1.1.0
msoffcrypto-tool==5.4.2
olefile==0.47
packaging==24.2
pillow==11.2.1
pluggy==1.6.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
psycopg2-binary==2.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
rarfile==4.2
redis==5.2.0
requests==2.32.3
requirements-parser==0.11.0
s3transfer==0.10.3
service-identity==24.2.0
setuptools==75.8.0
six==1.16.0
sqlparse==0.5.1
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
types-setuptools==80.4.0.20250511
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
wheel==0.37.1
zope.interface==7.2
